move shared pool to own file
